+++ src/components/layout/pageFooterPPT/PageFooterPPT.tsx
@@ -129,7 +129,10 @@ export default function PageFooterPPT({ pageNo, buttonDisabled = false }: PageFo
   };
 
   return (
-    <Paper sx={{ position: 'fixed', bottom: 40, left: 0, right: 0 }} elevation={0}>
+    <Paper
+      sx={{ backgroundColor: 'transparent', position: 'fixed', bottom: 40, left: 0, right: 0 }}
+      elevation={0}
+    >
       <Grid
         p={4}
         pt={0}
@@ -147,16 +150,18 @@ export default function PageFooterPPT({ pageNo, buttonDisabled = false }: PageFo
             />
           )}
         </Grid>
-        <Grid>
-          {(application.content5 as Notification)?.message?.length > 0 && (
-            <TimedAlert
-              color={(application.content5 as Notification)?.level}
-              delay={(application.content5 as Notification)?.delay}
-              testId="timeAlertFooter"
-              text={(application.content5 as Notification)?.message}
-            />
-          )}
-        </Grid>
+        {false && ( // Keep until it has been decided if we want to show messages here
+          <Grid>
+            {(application.content5 as Notification)?.message?.length > 0 && (
+              <TimedAlert
+                color={(application.content5 as Notification)?.level}
+                delay={(application.content5 as Notification)?.delay}
+                testId="timeAlertFooter"
+                text={(application.content5 as Notification)?.message}
+              />
+            )}
+          </Grid>
+        )}
         <Grid>
           {showNextNav() && (
             <NextPageButton