+++ tests/cypress/e2e/common/common.js
@@ -121,12 +121,12 @@ export const clickUserSearch = () => clickButton('userSearchButton');
 export const clickManageTeamMemberRights = () => clickButton('lockIcon');
 export const clickSubmitRights = () => clickButton('submitCheckbox');
 export const clickPICheckbox = () => clickButton('piCheckbox');
-export const clickAddProposal = () => clickButton('addProposalButton');
-export const clickAddMock = () => clickButton('addMockButton');
+export const clickAddProposal = () => clickButton('addSubmissionButton');
 export const clickCreateProposal = () => clickButton('nextButtonTestId');
 export const clickHome = () => clickButton('homeButtonTestId');
 export const clickDialogConfirm = () => clickButton('dialogConfirmationButton');
 export const clickLoginUser = () => clickButton('loginButton');
+export const clickCycleConfirm = () => clickButton('cycleConfirmationButton');
 export const clickUserMenu = () => clickButton('usernameMenu');
 export const clickObservationSetup = () => clickButton('addObservationButton');
 export const clickAddObservationEntry = () => clickButton('addObservationButtonEntry');
@@ -244,7 +244,7 @@ export const verifyUserMenuDecisions = exists => verifyUserMenu('menuItemReviewD
 /*----------------------------------------------------------------------*/
 
 export const pageConfirmed = label => cy.get('#pageTitle').contains(label);
-export const verifyOnLandingPage = () => verifyExists('addProposalButton');
+export const verifyOnLandingPage = () => verifyExists('addSubmissionButton');
 
 /*----------------------------------------------------------------------*/
 
@@ -292,12 +292,14 @@ export const createStandardProposal = () => {
 };
 
 export const createMock = () => {
-  clickAddMock();
+  clickAddProposal();
+  clickCycleConfirm();
   pageConfirmed('TARGET');
 };
 
 export const createStandardProposalLoggedIn = () => {
   clickAddProposal();
+  clickCycleConfirm();
   enterProposalTitle();
   clickProposalTypePrincipleInvestigator();
   clickSubProposalTypeTargetOfOpportunity();