+++ CHANGELOG.md
@@ -17,9 +17,11 @@ UNRELEASED
 * Changed: OSO-services : v5.1.0
 * Changed: PDF Viewer now responds to light / dark mode
 * Changed: Retrieval of OSD information now possible via utilities
+* Changed: Notifications moved into the standard footer
 * Added: Cycle dates to Landing and General pages
 * Changed: OSD constant data is combined with OSD API data into one accessor
 * Changed: Reduced direct usage of OSD constant data
+* Added: Cycle conformation when adding a submission
 
 5.1.0
 ----------